added a new model for "Hey GlaDOS"
it's horribly overfit with over a million steps on an RTX 3090 (~24 hours of total training time), with ~32k samples pre-augmenting. despite the sentence I just wrote, the model actually work surprisingly well. used the Jupyter notebook given in the microwakeword repository with some major edits.

target words:

- hey_glahdohs
- hey_glahdos
- hey_glahdows
- hey_glahdohss
- hey_glados
- hey_glaados
- hey_gladdos
- hey_gladoughs
- hey-glahdohs
- hey-glahdos
- hey-glahdows
- hey-glados
- hey-glaados
- hey-gladdos
- hey-gladoughs

noise-scales and noise-scale-ws was set to 0.5, min-phoneme-count was set to 10, positive class weight was 2, negative class weight was 100, and the minimization metric was set to ambient_false_positives_per_hour. the target minimization was set to 0.01 but according to the validation we got something much lower in practice.